+++ financius/src/main/java/com/code44/finance/ui/CalculatorActivity.java
@@ -6,12 +6,20 @@ import android.support.v4.app.Fragment;
 
 public class CalculatorActivity extends BaseActivity implements CalculatorFragment.CalculatorListener {
     public static final String RESULT_EXTRA_RESULT = "RESULT_EXTRA_RESULT";
+    public static final String RESULT_EXTRA_RAW_RESULT = "RESULT_EXTRA_RAW_RESULT";
 
     private static final String EXTRA_VALUE = "EXTRA_VALUE";
+    private static final String EXTRA_RAW_VALUE = "EXTRA_RAW_VALUE";
 
     public static void start(Fragment fragment, int requestCode, long value) {
         final Intent intent = makeIntent(fragment.getActivity(), CalculatorActivity.class);
         intent.putExtra(EXTRA_VALUE, value);
+        startForResult(fragment, intent, requestCode);
+    }
+
+    public static void start(Fragment fragment, int requestCode, double value) {
+        final Intent intent = makeIntent(fragment.getActivity(), CalculatorActivity.class);
+        intent.putExtra(EXTRA_RAW_VALUE, value);
         startForResult(fragment, intent, requestCode);
     }
 
@@ -21,18 +29,20 @@ public class CalculatorActivity extends BaseActivity implements CalculatorFragme
 
         // Get extras
         final long value = getIntent().getLongExtra(EXTRA_VALUE, 0);
-
+        final double rawValue = getIntent().getDoubleExtra(EXTRA_RAW_VALUE, 0);
 
         // Fragment
         if (savedInstanceState == null) {
-            getSupportFragmentManager().beginTransaction().add(android.R.id.content, CalculatorFragment.newInstance(value)).commit();
+            final CalculatorFragment fragment = getIntent().hasExtra(EXTRA_VALUE) ? CalculatorFragment.newInstance(value) : CalculatorFragment.newInstance(rawValue);
+            getSupportFragmentManager().beginTransaction().add(android.R.id.content, fragment).commit();
         }
     }
 
     @Override
-    public void onCalculatorResult(long result) {
+    public void onCalculatorResult(long result, double rawResult) {
         final Intent data = new Intent();
         data.putExtra(RESULT_EXTRA_RESULT, result);
+        data.putExtra(RESULT_EXTRA_RAW_RESULT, rawResult);
         setResult(RESULT_OK, data);
         finish();
     }